+++ src/scb/github.clj
@@ -2,7 +2,8 @@
   (:require
    [scb
     [utils :as utils]
-    [specs :as sp]]
+    
+    ]
    [clojure.spec.alpha :as s]
    [clojure.data.csv :as csv]
    [slugify.core :refer [slugify]]