+++ CHANGELOG.md
@@ -8,7 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-* github scraping
+* github catalogue support
+* tukui catalogue support
 * catalogue filtering and shortening
 * cli and an update script